move assistant mock to dedicated directory
