+++ functions/tests/_utils/mocking/platforms/assistant.js
