+++ functions/tests/actions/high-order-handlers/middlewares/feeder-from-playlist.spec.js
@@ -4,7 +4,7 @@ const rewire = require('rewire');
 const middleware = rewire('../../../../src/actions/high-order-handlers/middlewares/feeder-from-playlist');
 const playlist = require('../../../../src/state/playlist');
 
-const mockApp = require('../../../_utils/mocking/app');
+const mockApp = require('../../../_utils/mocking/platforms/assistant');
 const mockFeeders = require('../../../_utils/mocking/feeders');
 
 describe('actions', () => {
