+++ functions/tests/actions/high-order-handlers/middlewares/playback-fulfillment.spec.js
@@ -5,7 +5,7 @@ const middleware = rewire('../../../../src/actions/high-order-handlers/middlewar
 const playlist = require('../../../../src/state/playlist');
 const query = require('../../../../src/state/query');
 
-const mockApp = require('../../../_utils/mocking/app');
+const mockApp = require('../../../_utils/mocking/platforms/assistant');
 const mockDialog = require('../../../_utils/mocking/dialog');
 const mockFeeders = require('../../../_utils/mocking/feeders');
 const mockAlbumFeeder = require('../../../_utils/mocking/feeders/albums');
