+++ functions/tests/actions/media-status-update.spec.js
@@ -5,7 +5,7 @@ const action = rewire('../../src/actions/media-status-update');
 const playlist = require('../../src/state/playlist');
 const query = require('../../src/state/query');
 
-const mockApp = require('../_utils/mocking/app');
+const mockApp = require('../_utils/mocking/platforms/assistant');
 const mockDialog = require('../_utils/mocking/dialog');
 const mockMiddlewares = require('../_utils/mocking/middlewares');
