+++ functions/tests/platform/assistant/persistence/session.spec.js
@@ -1,6 +1,6 @@
 const {expect} = require('chai');
 
-const mockAssistant = require('../../../_utils/mocking/app');
+const mockAssistant = require('../../../_utils/mocking/platforms/assistant');
 const persistance = require('../../../../src/platform/assistant/persistence/session');
 
 describe('platform', () => {
