+++ functions/tests/state/playlist.spec.js
@@ -2,7 +2,7 @@ const {expect} = require('chai');
 
 const playlist = require('../../src/state/playlist');
 
-const mockApp = require('../_utils/mocking/app');
+const mockApp = require('../_utils/mocking/platforms/assistant');
 
 describe('playlist', () => {
   let app;
